Feat - Badge Generator (#59)
* chore: Sketch out initial badge

This needs some UI work so we can test the variants. Also the template
needs to be expanded and backend schema is needed so this all can be
tied together.

* chore(badges): Set up a datatype for Attendee

* chore(badges): Use full names for templates + import reflection to app

Reflection is needed by type-graphql. Strangely enough the related types
worked before. Maybe it's something in the Attendee model or so.

* chore: Shim type-graphql

No need to bring it to the client. We consume just the TS types.

* refactor(asset-designer): Generalize templateId to selectionId

This id is needed for tracking between templates/components (assumed to
be unique).

* feat(asset-designer): Allow Badge component to be selected

* chore(asset-designer): Prototype class field reflection

This needs more work but now the hard part seems to be done. Still need
to change this to use the same data format as variables.

* chore: Drop a redundant console.log

Better for demoing without.

* chore(badge): Sketch out variables

* wip: Sketch out nested component field generation

* fix(asset-designer): Fix nested field keying

* chore: Render variable fields in a nicer manner

Now it looks better.

* fix(asset-designer): Render select for enums

Still missing some wiring (props, router).

* chore: Sketch out selected/variables parsing in backend

Still need to fix the frontend bit.

* chore: Sketch out how query validation should work

Still need to push a query to frontend. That also allows me to avoid the
last remaining page refresh after the app is running.

* chore(asset-designer): Simplify code

Better now.

* chore: Push themes query to frontend

Simpler this way although it's extra roundtrip (doesn't matter at this
point). Likely there's a good way to hydrate the initial data anyway
without this complexity.

* fix(asset-designer): Fix UI for selects

Still need to check the resulting queries.

* fix(header-template): Pick up graphql context from a prop

Better this way.

* chore(asset-designer): Simplify code

* fix(asset-designer): Fix schedule queries

* fix(asset-designer): Fix speaker tweet queries

* refactor(asset-designer): Drop the concept of "selected"

That was merged into variables.

* fix(asset-designer): Fix HeaderTemplate rendering

No more selected.

* fix(asset-designer): Fix query encoding

Better encode only the object structure.

* fix(asset-designer): Pick up template id from the query

* chore: Sketch out a template for business cards

* chore(business-card): Improve layout

* chore: Sketch out two-sided business card

The idea will be needed for badges too.

* feat: Allow business card to be rotated on how

There's also a toggle control for this so it works without hover too.
Still need to generalize a bit and port also to badges.

* refactor(asset-designer): Simplify router logic

No need to validate as that can crash hard. It's better to pass through
to keep some kind of backwards compatibility.

* chore: Document Field shim better

* feat: Populate complex type defaults

* Revert "refactor(asset-designer): Simplify router logic"

This reverts commit 12cb70a2e7a7a9dd1882e2d7f05dd43319224b44.

* chore: Update package lock

* refactor: Move the handling of initial variables to frontend

This also allows me to drop backend validation step properly. Previously
it was coupled with the handling of defaults.

* fix(asset-designer): Make variable fields work for nested cases

* refactor(asset-designer): Separate VariableSelector

* refactor(asset-designer): Extract AssetDesignerSidebar

* chore(badge): Add color customization

* refactor: Extract texture to Theme model

* fix(asset-designer): Fix asset designer page

* fix: Remove fixed texture references

Now it should be dynamic.

* chore: Wire new badges together with pages

This needs some more work styling-wise.

* feat(badge): Set up initial print CSS

Also made asset designer a bit more robust (some lookup weirdness).
Still need to add other sides for the badges and then this is done.

* chore: Add a todo

* feat(badges): Render backsides for the print

* refactor(badges): Clean up the way empty badges are generated

This needs to be exposed to the UI still. That requires some thought.

* fix(variable-selector): Cast numbers to Number onChange

onChange returns strings.

* feat(badges): Expose fields for empties

This can likely be done better but it needs dependent types (i.e. field
definition and then empty fields based on that). A little trickier.

* feat: Implement initial version of the attendee api

This looks into the filesystem for a csv file. If it exists, you'll get
results.

Still need to connect this to the badge generator.

* feat(badges): Connect to attendee data if it exists

Likely `connect` should be cleaned up (needs a post-process for getting
data).